Extract ValidatedServerConfig
